Consolidate `Error` types for parsers
